# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Atlas - Project-based technical learning platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects with optional filters
      parameters:
        - name: domainSlug
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [beginner, intermediate, advanced, expert]
        - name: tier
          in: query
          schema:
            type: string
            enum: [free, pro]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedProjects"
  /projects/{slug}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project details including steps
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project detail with steps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    ProjectSummary:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced, expert]
        tier:
          type: string
          enum: [free, pro]
        xpReward:
          type: integer
        estimatedHours:
          type: number
        stepCount:
          type: integer
        enrolledCount:
          type: integer
        completionRate:
          type: number
        tags:
          type: array
          items:
            type: string
        position:
          type: integer
      required:
        - id
        - slug
        - title
        - description
        - difficulty
        - tier
        - xpReward
        - estimatedHours
        - stepCount
        - enrolledCount
        - completionRate
        - tags
        - position
    PaginatedProjects:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProjectSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        hasMore:
          type: boolean
      required:
        - data
        - total
        - page
        - limit
        - hasMore
    ProjectStep:
      type: object
      properties:
        id:
          type: string
        position:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [code_python, code_sql, multiple_choice, true_false, short_answer, file_upload, diagram, concept_check, reflection]
        starterCode:
          type: string
        expectedOutput:
          type: string
        hints:
          type: array
          items:
            type: string
        xpReward:
          type: integer
        isLocked:
          type: boolean
      required:
        - id
        - position
        - title
        - description
        - type
        - xpReward
        - isLocked
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/ProjectSummary"
        - type: object
          properties:
            longDescription:
              type: string
            learningObjectives:
              type: array
              items:
                type: string
            prerequisites:
              type: array
              items:
                type: string
            steps:
              type: array
              items:
                $ref: "#/components/schemas/ProjectStep"
            domainSlug:
              type: string
            domainName:
              type: string
...